requirements.txt
APScheduler==3.10.1
asgiref==3.6.0
certifi==2023.7.22
channels==4.0.0
charset-normalizer==3.3.0
crispy-bootstrap4==2022.1
cryptography==41.0.4
defusedxml==0.7.1
dj-database-url==1.2.0
Django==4.1.7
django-allauth==0.57.0
django-apscheduler==0.6.2
django-braces==1.15.0
django-cors-headers==4.3.1
django-countries==7.5.1
django-crispy-forms==2.0
django-crontab==0.7.1
django-filter==23.3
django-model-utils==4.3.1
django-phone-auth==0.3.1
django-phonenumber-field==7.0.2
django-phonenumbers==1.0.1
django-rest-auth==0.9.5
django-star-ratings==0.9.2
django_rest_resetpassword==0.1.2
djangorestframework==3.15.1
djangorestframework-jwt==1.11.0
djangorestframework-simplejwt==5.3.0
gunicorn==20.1.0
idna==3.4
Markdown==3.5
numpy==1.26.4
oauthlib==3.2.2
pandas==2.2.1
pycparser==2.21
PyJWT==1.7.1
python-dateutil==2.9.0.post0
python3-openid==3.2.0
PythonTurtle==0.3.2
pytz==2022.7.1
pytz-deprecation-shim==0.1.0.post0
PyYAML==6.0.1
requests==2.31.0
schedule==1.1.0
six==1.16.0
sqlparse==0.4.3
swapper==1.3.0
tk==0.1.0
typing_extensions==4.5.0
tzdata==2022.7
tzlocal==4.3
urllib3==2.0.6
whitenoise==6.4.0
